omicstoken_repo/requirements.txt
fastapi
uvicorn
python-multipart
pandas
numpy
faiss-cpu
google-generativeai>=0.3.0
httpx
